feat: added view for other frameworks
